+++ CHANGELOG.md
@@ -15,8 +15,8 @@ CHANGELOG
 * The `DatabaseReader` methods `city()` and `country()` can now be called on
   the Enterprise database and the `country()` method can be called on City
   databases. Request by Gergely Boromissza. GitHub #132.
-* A new getStaticIPScore method was added to `com.maxmind.geoip2.record.Traits`
-  for use with GeoIP2 Precision Insights.
+* New `getStaticIPScore()` and `getUserCount()` methods were added to
+  `com.maxmind.geoip2.record.Traits` for use with GeoIP2 Precision Insights.
 
 2.12.0 (2018-04-11)
 -------------------